+++ lib/rdf/data.ex
@@ -92,6 +92,11 @@ defprotocol RDF.Data do
   """
   def statement_count(data)
 
+  @doc """
+  Returns a nested map of the native Elixir values of a RDF data structure.
+  """
+  def values(data)
+
 end
 
 defimpl RDF.Data, for: RDF.Description do
@@ -158,6 +163,7 @@ defimpl RDF.Data, for: RDF.Description do
 
   def subject_count(_),             do: 1
   def statement_count(description), do: RDF.Description.count(description)
+  def values(description),          do: RDF.Description.values(description)
 end
 
 
@@ -218,6 +224,7 @@ defimpl RDF.Data, for: RDF.Graph do
 
   def subject_count(graph),   do: RDF.Graph.subject_count(graph)
   def statement_count(graph), do: RDF.Graph.triple_count(graph)
+  def values(graph),          do: RDF.Graph.values(graph)
 end
 
 
@@ -271,4 +278,5 @@ defimpl RDF.Data, for: RDF.Dataset do
 
   def subject_count(dataset),   do: dataset |> subjects |> Enum.count
   def statement_count(dataset), do: RDF.Dataset.statement_count(dataset)
+  def values(dataset),          do: RDF.Dataset.values(dataset)
 end